Slack_automation/pyproject.toml
[project]
name = "slack-automation-of-html-reports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "slack-sdk>=3.39.0",
]
